projects/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
py-algorand-sdk==2.7.0
google-generativeai==0.8.4
Pillow==11.1.0
